factorized and decoupled a lot of code to resolve bug
